Update imagingTransaction.xlsx
Added some data validation to help stop the spreadsheet data from being deleted. It will prevent the user from changeing a row in which the due date is past, stop from allowing to be set to nothing, and doesn't allow negative numbers/letters.

The only upkeep is extending the formating and formulas when the sheet is full.